refactor: organize plaid specifics
