+++ src/almonds/db/base.py
@@ -10,5 +10,5 @@ from almonds.models.category import Category
 from almonds.models.goal import Goal
 from almonds.models.transaction import Transaction
 from almonds.models.user import User
-from almonds.models.plaid_item import PlaidItem
+from almonds.models.plaid.plaid_item import PlaidItem
 from almonds.models.user_settings import UserSetting
